split predict.models into .models and .base
