+++ glasspy/predict/__init__.py
@@ -1,0 +1,7 @@
+from .models import ViscNet
+
+__all__ = [
+    ViscNet,
+    ViscNetHuber,
+    ViscNetVFT,
+]
